+++ lib/src/constants.dart
@@ -7,10 +7,11 @@ class TurnColor {
 }
 
 class MoveMark {
-  /// PASS string.
-  ///
-  /// NOTE: edaxGetMoves() can contain passString upper case `PA`.
-  static const passString = 'pa';
+  /// PASS string of BLACK.
+  static const passStringOfBlack = 'PA';
+
+  /// PASS string of WHITE.
+  static const passStringOfWhite = 'pa';
 
   /// PASS move.
   static const pass = 64;