+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 0.10.0-beta.0
 
-- add `computeBestPathNumWithLink`
+- **Breaking Changes**
+  - fix `PASS` String.
+    - separate `PA (BLACK)` and `pa (WHITE)`
+- New Features
+  - add `computeBestPathNumWithLink`
 
 # 0.9.0-beta.0